#10334: Migrate angle_bw to TTNN
