+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -637,8 +637,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.complex_abs_bw
 
-.. autofunction:: tt_lib.tensor.angle_bw
-
 .. autofunction:: tt_lib.tensor.complex_recip_bw
 
 .. autofunction:: tt_lib.tensor.complex_mul_bw
