+++ docs/source/ttnn/ttnn/ttnn/angle_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.angle_bw:
+
+ttnn.angle_bw
+###################
+
+.. autofunction:: ttnn.angle_bw
